Added some comments, split Editor and Item components into separate files. Some other refactoring.
